+++ src/System.IO.FileSystem.Watcher.Polling/System/IO/FileState.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Diagnostics;
+
+namespace System.IO.FileSystem
+{
+    [Serializable]
+    struct FileState
+    {
+        internal long _version;  // removal notification are implemented something similar to "mark and sweep". This value is incremented in the mark phase
+        public string Path;
+        public string Directory;
+        public DateTimeOffset LastWriteTimeUtc;
+        public long Length;
+
+        public FileState(string directory, string path) : this()
+        {
+            Debug.Assert(path != null);
+            Directory = directory;
+            Path = path;
+        }
+
+        public override string ToString()
+        {
+            return Path;
+        }
+    }
+}
